Move the logic to sidebar state
