+++ core-ui/src/state/navigation/extensionsAtom.ts
@@ -1,16 +1,11 @@
 import jsyaml from 'js-yaml';
-import { mapValues, partial } from 'lodash';
+import { mapValues } from 'lodash';
 import { useEffect } from 'react';
 import { ExtResource } from '../types';
 import { atom, RecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
 import { clusterState } from '../clusterAtom';
 import { authDataState } from '../authDataAtom';
 import { getFetchFn } from '../utils/getFetchFn';
-import { configurationAtom } from 'state/configuration/configurationAtom';
-import { openapiPathIdListSelector } from 'state/openapi/openapiPathIdSelector';
-import { permissionSetsSelector } from 'state/permissionSetsSelector';
-import { shouldNodeBeVisible } from './filters/shouldNodeBeVisible';
-import { mapExtResourceToNavNode } from 'state/resourceList/mapExtResourceToNavNode';
 
 type ConfigMapData = {
   general: string;
@@ -89,10 +84,6 @@ export const useGetExtensions = () => {
   const fetchFn = getFetchFn(useRecoilValue);
   const cluster = useRecoilValue(clusterState);
   const auth = useRecoilValue(authDataState);
-  const configuration = useRecoilValue(configurationAtom);
-  const features = configuration?.features;
-  const openapiPathIdList = useRecoilValue(openapiPathIdListSelector);
-  const permissionSet = useRecoilValue(permissionSetsSelector);
 
   useEffect(() => {
     const manageExtensions = async () => {
@@ -100,23 +91,7 @@ export const useGetExtensions = () => {
         setExtensions(null);
       } else {
         const configs = await getExtensions(fetchFn);
-        if (!configs) {
-          setExtensions(null);
-        } else {
-          const configSet = {
-            configFeatures: features!,
-            openapiPathIdList,
-            permissionSet,
-          };
-          const isNodeVisibleForCurrentConfigSet = partial(
-            shouldNodeBeVisible,
-            configSet,
-          );
-          const filteredConfigs = configs.filter(node =>
-            isNodeVisibleForCurrentConfigSet(mapExtResourceToNavNode(node)),
-          );
-          setExtensions(filteredConfigs);
-        }
+        setExtensions(configs);
       }
     };
     manageExtensions();
